Revert "Make QueryLoader messages appear inside tx table"
This reverts commit 5c6ab16827e7a8e27e7a0a0143e643afde41e365.
