+++ src/pages/Admin/Charity/Withdraws/Transactions/index.tsx
@@ -1,11 +1,28 @@
+import { useAdminResources } from "pages/Admin/Guard";
+import { useWithdrawLogsQuery } from "services/apes";
+import QueryLoader from "components/QueryLoader";
 import List from "./List";
 import Table from "./Table";
 
 export default function Transactions() {
+  const { cw3 } = useAdminResources();
+  const queryState = useWithdrawLogsQuery(cw3);
   return (
-    <>
-      <List classes="grid md:hidden" />
-      <Table classes="hidden md:table" />
-    </>
+    <QueryLoader
+      queryState={queryState}
+      messages={{
+        loading: "Loading transactions...",
+        error: "Failed to get transactions",
+        empty: "No transactions found",
+      }}
+    >
+      {(logs) => (
+        <>
+          <List withdraws={logs} classes="grid md:hidden" />
+          <Table withdraws={logs} classes="hidden md:table" />
+        </>
+      )}
+    </QueryLoader>
   );
+  // return <Table withdraws={logs} classes="hidden md:table" />;
 }
